requirements.txt
# ─────────────────────────────────────────────
#  Climate AI v2 — Production Requirements
# ─────────────────────────────────────────────

# Google Earth Engine
earthengine-api>=0.1.391
geemap>=0.32.0

# Data Storage & Processing
duckdb>=0.10.2
pandas>=2.2.3
numpy>=1.26.4
xarray>=2024.3.0
netCDF4>=1.6.5,!=1.7.1
geopandas>=0.14.4
shapely>=2.0.4
pyproj>=3.6.1

# Machine Learning
tensorflow>=2.16.1
scikit-learn>=1.4.2
mlflow>=2.12.2
joblib>=1.4.2
setuptools<70.0.0
shap>=0.45.0
scipy>=1.13.0
geopy>=2.4.1

# Backend API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.1
python-multipart>=0.0.9
httpx>=0.27.0

# Frontend
streamlit>=1.33.0
pydeck>=0.9.1
plotly>=5.21.0
folium>=0.16.0
streamlit-folium>=0.20.0

# Utils
python-dotenv>=1.0.1
loguru>=0.7.2
tqdm>=4.66.4
requests>=2.31.0
apscheduler>=3.10.4
